Test.Hspec.Core.Extension.Item

Description

Warning: This API is experimental.

Types

data Item a

Item is used to represent spec items internally. A spec item consists of:

a textual description of a desired behavior

an example for that behavior

additional meta information

Everything that is an instance of the Example type class can be used as an example, including QuickCheck properties, Hspec expectations and HUnit assertions.

Constructors

Item

Fields

itemRequirement :: String

Textual description of behavior

itemLocation :: Maybe Location

Source location of the spec item

itemIsParallelizable :: Maybe Bool

A flag that indicates whether it is safe to evaluate this spec item in parallel with other spec items

itemIsFocused :: Bool

A flag that indicates whether this spec item is focused

itemAnnotations :: Annotations

Arbitrary additional data that can be used by third-party extensions.

Since: 2.12.0

itemExample :: Params -> (ActionWith a -> IO ()) -> ProgressCallback -> IO Result

Example for behavior

data Location

Location is used to represent source locations.

Constructors

Location

Fields

locationFile :: FilePath

locationLine :: Int

locationColumn :: Int

data Params

QuickCheck and SmallCheck related parameters.

Constructors

Params

Fields

paramsQuickCheckArgs :: Args

paramsSmallCheckDepth :: Maybe Int

type ActionWith a = a -> IO ()

An IO action that expects an argument of type a.

This type is what Examples are ultimately unlifted into for execution.

type Progress = (Int, Int)

(CurrentItem, TotalItems) tuple.

type ProgressCallback = Progress -> IO ()

Callback used by composite test items that contain many tests to report their progress towards finishing them all.

This is used, for example, to report how many QuickCheck examples are finished.

data Result

The result of running an example

Constructors

Result

Fields

resultInfo :: String

resultStatus :: ResultStatus

data ResultStatus

Constructors

Success

Pending (Maybe Location) (Maybe String)

Failure (Maybe Location) FailureReason

data FailureReason

Constructors

NoReason

Reason String

ColorizedReason String

ExpectedButGot (Maybe String) String String

Error (Maybe String) SomeException

Operations

isFocused :: Item a -> Bool

pending :: Item a -> Item a

pendingWith :: String -> Item a -> Item a

setAnnotation :: Typeable value => value -> Item a -> Item a

getAnnotation :: Typeable value => Item a -> Maybe value