Test.Hspec.Core.Format

Description

This is an unstable API. Use Test.Hspec.Api.Format.V2 instead.

Documentation

type Format = Event -> IO ()

data FormatConfig

Since: 2.11.5

Constructors

FormatConfig

Fields

formatConfigUseColor :: Bool

formatConfigReportProgress :: Bool

formatConfigOutputUnicode :: Bool

formatConfigUseDiff :: Bool

formatConfigDiffContext :: Maybe Int

formatConfigExternalDiff :: Maybe (String -> String -> IO ())

formatConfigPrettyPrint :: Bool

Deprecated: Use formatConfigPrettyPrintFunction instead

formatConfigPrettyPrintFunction :: Maybe (String -> String -> (String, String))

formatConfigFormatException :: SomeException -> String

Since: 2.11.5

formatConfigPrintTimes :: Bool

formatConfigHtmlOutput :: Bool

formatConfigPrintCpuTime :: Bool

formatConfigUsedSeed :: Integer

formatConfigExpectedTotalCount :: Int

formatConfigExpertMode :: Bool

defaultFormatConfig :: FormatConfig

data Event

Constructors

Started

GroupStarted Path

GroupDone Path

Progress Path Progress

ItemStarted Path

ItemDone Path Item

Done [(Path, Item)]

type Progress = (Int, Int)

type Path = ([String], String)

A Path describes the location of a spec item within a spec tree.

It consists of a list of group descriptions and a requirement description.

Since: 2.0.0

data Location

Location is used to represent source locations.

Constructors

Location

Fields

locationFile :: FilePath

locationLine :: Int

locationColumn :: Int

newtype Seconds

Constructors

Seconds Double

data Item

Constructors

Item

Fields

itemLocation :: Maybe Location

itemDuration :: Seconds

itemInfo :: String

itemResult :: Result

data Result

Constructors

Success

Pending (Maybe Location) (Maybe String)

Failure (Maybe Location) FailureReason

data FailureReason

Constructors

NoReason

Reason String

ColorizedReason String

ExpectedButGot (Maybe String) String String

Error (Maybe String) SomeException

monadic :: MonadIO m => (m () -> IO ()) -> (Event -> m ()) -> IO Format