Test.Hspec.Core.Formatters.V2

Description

This is an unstable API. Use Test.Hspec.Api.Formatters.V3 instead.

Formatters

silent :: Formatter

checks :: Formatter

specdoc :: Formatter

progress :: Formatter

failed_examples :: Formatter

Implementing a custom Formatter

A formatter is a set of actions. Each action is evaluated when a certain situation is encountered during a test run.

Actions live in the FormatM monad. It provides access to the runner state and primitives for appending to the generated report.

data Formatter

Constructors

Formatter

Fields

formatterStarted :: FormatM ()

formatterGroupStarted :: Path -> FormatM ()

formatterGroupDone :: Path -> FormatM ()

formatterProgress :: Path -> Progress -> FormatM ()

formatterItemStarted :: Path -> FormatM ()

formatterItemDone :: Path -> Item -> FormatM ()

formatterDone :: FormatM ()

type Path = ([String], String)

A Path describes the location of a spec item within a spec tree.

It consists of a list of group descriptions and a requirement description.

Since: 2.0.0

type Progress = (Int, Int)

data Location

Location is used to represent source locations.

Constructors

Location

Fields

locationFile :: FilePath

locationLine :: Int

locationColumn :: Int

data Item

Constructors

Item

Fields

itemLocation :: Maybe Location

itemDuration :: Seconds

itemInfo :: String

itemResult :: Result

data Result

Constructors

Success

Pending (Maybe Location) (Maybe String)

Failure (Maybe Location) FailureReason

data FailureReason

Constructors

NoReason

Reason String

ColorizedReason String

ExpectedButGot (Maybe String) String String

Error (Maybe String) SomeException

data FormatM a

formatterToFormat :: Formatter -> FormatConfig -> IO Format

Accessing config values

getConfig :: FormatM FormatConfig

Since: 2.11.5

getConfigValue :: (FormatConfig -> a) -> FormatM a

Since: 2.11.5

data FormatConfig

Since: 2.11.5

Constructors

FormatConfig

Fields

formatConfigUseColor :: Bool

formatConfigReportProgress :: Bool

formatConfigOutputUnicode :: Bool

formatConfigUseDiff :: Bool

formatConfigDiffContext :: Maybe Int

formatConfigExternalDiff :: Maybe (String -> String -> IO ())

formatConfigPrettyPrint :: Bool

Deprecated: Use formatConfigPrettyPrintFunction instead

formatConfigPrettyPrintFunction :: Maybe (String -> String -> (String, String))

formatConfigFormatException :: SomeException -> String

Since: 2.11.5

formatConfigPrintTimes :: Bool

formatConfigHtmlOutput :: Bool

formatConfigPrintCpuTime :: Bool

formatConfigUsedSeed :: Integer

formatConfigExpectedTotalCount :: Int

formatConfigExpertMode :: Bool

Accessing the runner state

getSuccessCount :: FormatM Int

Get the number of successful examples encountered so far.

getPendingCount :: FormatM Int

Get the number of pending examples encountered so far.

getFailCount :: FormatM Int

Get the number of failed examples encountered so far.

getTotalCount :: FormatM Int

Get the total number of examples encountered so far.

getExpectedTotalCount :: FormatM Int

Get the number of spec items that will have been encountered when this run completes (if it is not terminated early).

Since: 2.9.0

data FailureRecord

Constructors

FailureRecord

Fields

failureRecordLocation :: Maybe Location

failureRecordPath :: Path

failureRecordMessage :: FailureReason

getFailMessages :: FormatM [FailureRecord]

Get the list of accumulated failure messages.

usedSeed :: FormatM Integer

The random seed that is used for QuickCheck.

printTimes :: FormatM Bool

Return True if the user requested time reporting for individual spec items, False otherwise.

newtype Seconds

Constructors

Seconds Double

getCPUTime :: FormatM (Maybe Seconds)

Get the used CPU time since the test run has been started.

getRealTime :: FormatM Seconds

Get the passed real time since the test run has been started.

Appending to the generated report

write :: String -> FormatM ()

Append some output to the report.

writeLine :: String -> FormatM ()

The same as write, but adds a newline character.

writeTransient :: String -> FormatM ()

Dealing with colors

withInfoColor :: FormatM a -> FormatM a

Set output color to cyan, run given action, and finally restore the default color.

withSuccessColor :: FormatM a -> FormatM a

Set output color to green, run given action, and finally restore the default color.

withPendingColor :: FormatM a -> FormatM a

Set output color to yellow, run given action, and finally restore the default color.

withFailColor :: FormatM a -> FormatM a

Set output color to red, run given action, and finally restore the default color.

outputUnicode :: FormatM Bool

Return True if the user requested unicode output, False otherwise.

Since: 2.9.0

useDiff :: FormatM Bool

Return True if the user requested colorized diffs, False otherwise.

diffContext :: FormatM (Maybe Int)

Return the value of configDiffContext.

Since: 2.10.6

externalDiffAction :: FormatM (Maybe (String -> String -> IO ()))

An action for printing diffs.

The action takes expected and actual as arguments.

When this is a Just-value then it should be used instead of any built-in diff implementation. A Just-value also implies that useDiff returns True.

Since: 2.10.6

prettyPrint :: FormatM Bool

Deprecated: use prettyPrintFunction instead

Return True if the user requested pretty diffs, False otherwise.

prettyPrintFunction :: FormatM (Maybe (String -> String -> (String, String)))

Return a function for pretty-printing if the user requested pretty diffs, Nothing otherwise.

Since: 2.10.0

extraChunk :: String -> FormatM ()

Output given chunk in red.

missingChunk :: String -> FormatM ()

Output given chunk in green.

expert mode

unlessExpert :: FormatM () -> FormatM ()

Do nothing on `--expert`, otherwise run the given action.

Since: 2.11.2

Helpers

formatLocation :: Location -> String

formatException :: SomeException -> String

The function formatException converts an exception to a string.

This is different from show. The type of the exception is included, e.g.:

>>> formatException (toException DivideByZero)
"ArithException\ndivide by zero"

For IOExceptions the IOErrorType is included, as well.

Since: 2.0.0