Test.Hspec.Core.Spec

Description

This module provides access to Hspec's internals. It is less stable than other parts of the API. For most users Test.Hspec is more suitable!

Defining a spec

it :: (HasCallStack, Example a) => String -> a -> SpecWith (Arg a)

The it function creates a spec item.

A spec item consists of:

a textual description of a desired behavior

an example for that behavior

describe "absolute" $ do
  it "returns a positive number when given a negative number" $
    absolute (-1) == 1

Example a optionally accepts an argument Arg a, which is then given to the test body. This is useful for provisioning resources for a test which are created and cleaned up outside the test itself. See Arg for details.

Note that this function is often on the scene of nasty type errors due to GHC failing to infer the type of do notation in the test body. It can be helpful to use TypeApplications to explicitly specify the intended Example type.

specify :: (HasCallStack, Example a) => String -> a -> SpecWith (Arg a)

specify is an alias for it.

describe :: HasCallStack => String -> SpecWith a -> SpecWith a

The describe function combines a list of specs into a larger spec.

context :: HasCallStack => String -> SpecWith a -> SpecWith a

context is an alias for describe.

pending :: HasCallStack => Expectation

pending can be used to mark a spec item as pending.

If you want to textually specify a behavior but do not have an example yet, use this:

describe "fancyFormatter" $ do
  it "can format text in a way that everyone likes" $
    pending

pendingWith :: HasCallStack => String -> Expectation

pendingWith is similar to pending, but it takes an additional string argument that can be used to specify the reason for why the spec item is pending.

xit :: (HasCallStack, Example a) => String -> a -> SpecWith (Arg a)

Changing it to xit marks the corresponding spec item as pending.

This can be used to temporarily disable a spec item.

xspecify :: (HasCallStack, Example a) => String -> a -> SpecWith (Arg a)

xspecify is an alias for xit.

xdescribe :: HasCallStack => String -> SpecWith a -> SpecWith a

Changing describe to xdescribe marks all spec items of the corresponding subtree as pending.

This can be used to temporarily disable spec items.

xcontext :: HasCallStack => String -> SpecWith a -> SpecWith a

xcontext is an alias for xdescribe.

Focused spec items

During a test run, when a spec contains focused spec items, all other spec items are ignored.

focus :: SpecWith a -> SpecWith a

focus focuses all spec items of the given spec.

Applying focus to a spec with focused spec items has no effect.

fit :: (HasCallStack, Example a) => String -> a -> SpecWith (Arg a)

fit is an alias for fmap focus . it

fspecify :: (HasCallStack, Example a) => String -> a -> SpecWith (Arg a)

fspecify is an alias for fit.

fdescribe :: HasCallStack => String -> SpecWith a -> SpecWith a

fdescribe is an alias for fmap focus . describe

fcontext :: HasCallStack => String -> SpecWith a -> SpecWith a

fcontext is an alias for fdescribe.

parallel :: SpecWith a -> SpecWith a

parallel marks all spec items of the given spec to be safe for parallel evaluation.

sequential :: SpecWith a -> SpecWith a

sequential marks all spec items of the given spec to be evaluated sequentially.

The SpecM monad

type Spec = SpecWith ()

A SpecWith that can be evaluated directly by the hspec function as it does not require any parameters.

type SpecWith a = SpecM a ()

A SpecWith a represents a test or group of tests that require an a value to run.

In the common case, a Spec is a SpecWith () which requires () and can thus be executed with hspec.

To supply an argument to SpecWith tests to turn them into Spec, use functions from Test.Hspec.Core.Hooks such as around, before, mapSubject and similar.

Values of this type are created by it, describe and similar.

newtype SpecM a r

A writer monad for SpecTree forests.

This is used by describe and is used to construct the forest of spec items.

It allows for dynamically generated spec trees, for example, by using data obtained by performing IO actions with runIO.

Constructors

SpecM (WriterT (Endo Config, [SpecTree a]) (ReaderT Env IO) r)

runSpecM :: SpecWith a -> IO (Endo Config, [SpecTree a])

Convert a Spec to a forest of SpecTrees.

fromSpecList :: [SpecTree a] -> SpecWith a

Create a Spec from a forest of SpecTrees.

runIO :: IO r -> SpecM a r

Run an IO action while constructing the spec tree.

SpecM is a monad to construct a spec tree, without executing any spec items itself. runIO allows you to run IO actions during this construction phase. The IO action is always run when the spec tree is constructed (e.g. even when --dry-run is specified). If you do not need the result of the IO action to construct the spec tree, beforeAll may be more suitable for your use case.

mapSpecForest :: ([SpecTree a] -> [SpecTree b]) -> SpecM a r -> SpecM b r

mapSpecItem :: (ActionWith a -> ActionWith b) -> (Item a -> Item b) -> SpecWith a -> SpecWith b

Deprecated: Use mapSpecItem_ instead.

mapSpecItem_ :: (Item a -> Item b) -> SpecWith a -> SpecWith b

modifyParams :: (Params -> Params) -> SpecWith a -> SpecWith a

Modifies the Params on the spec items to be generated by SpecWith.

modifyConfig :: (Config -> Config) -> SpecWith a

Since: 2.10.0

getSpecDescriptionPath :: SpecM a [String]

Get the path of describe labels, from the root all the way in to the call-site of this function.

Example

Expand

>>> :{
runSpecM $ do
  describe "foo" $ do
    describe "bar" $ do
      getSpecDescriptionPath >>= runIO . print
:}
["foo","bar"]

Since: 2.10.0

A type class for examples

class Example e where

A type class for examples, that is to say, test bodies as used in it and similar functions.

Associated Types

type Arg e

The argument type that is needed to run this Example. If Arg is (), no argument is required and the Example can be run as-is.

The value of Arg is the difference between Spec (aka SpecWith ()), which can be executed, and SpecWith a, which cannot be executed without turning it into Spec first.

To supply an argument to examples, use the functions in Test.Hspec.Core.Hooks such as around, before, mapSubject and similar.

type Arg e = ()

Methods

evaluateExample

Arguments

:: e
The example being evaluated

-> Params
QuickCheck/SmallCheck settings

-> (ActionWith (Arg e) -> IO ())
Hook: takes an ActionWith (Arg e), namely, the IO action to run the test body, obtains Arg e from somewhere, then executes the test body (or possibly decides not to execute it!).

This is used to implement around and similar hook functionality.

-> ProgressCallback
Callback for composite tests like QuickCheck to report their progress.

-> IO Result

Evaluates an example.

evaluateExample is expected to execute the test body inside the IO action passed to the hook. It's often necessary to use an IORef to pass data out like whether the test succeeded to the outer IO action so it can be returned as a Result.

Example:

newtype MyAction = MyAction (Int -> IO Bool)

instance Example MyAction where
  type Arg MyAction = Int

  evaluateExample (MyAction act) _params hook _progress = do
    result <- newIORef (Result "" Success)
    hook $ arg -> do
      -- e.g. determines if arg is 42
      ok <- act arg
      let result' = Result "" $ if ok then Success else Failure Nothing NoReason
      writeIORef result result'
    readIORef result

data Params

QuickCheck and SmallCheck related parameters.

Constructors

Params

Fields

paramsQuickCheckArgs :: Args

paramsSmallCheckDepth :: Maybe Int

defaultParams :: Params

type ActionWith a = a -> IO ()

An IO action that expects an argument of type a.

This type is what Examples are ultimately unlifted into for execution.

type Progress = (Int, Int)

(CurrentItem, TotalItems) tuple.

type ProgressCallback = Progress -> IO ()

Callback used by composite test items that contain many tests to report their progress towards finishing them all.

This is used, for example, to report how many QuickCheck examples are finished.

data Result

The result of running an example

Constructors

Result

Fields

resultInfo :: String

resultStatus :: ResultStatus

data ResultStatus

Constructors

Success

Pending (Maybe Location) (Maybe String)

Failure (Maybe Location) FailureReason

data Location

Location is used to represent source locations.

Constructors

Location

Fields

locationFile :: FilePath

locationLine :: Int

locationColumn :: Int

data FailureReason

Constructors

NoReason

Reason String

ColorizedReason String

ExpectedButGot (Maybe String) String String

Error (Maybe String) SomeException

safeEvaluate :: IO Result -> IO Result

safeEvaluateExample :: Example e => e -> Params -> (ActionWith (Arg e) -> IO ()) -> ProgressCallback -> IO Result

Internal representation of a spec tree

type SpecTree a = Tree (IO ()) (Item a)

A tree is used to represent a spec internally. The tree is parameterized over the type of cleanup actions and the type of the actual spec items.

data Tree c a

Internal tree data structure

Constructors

Node String [Tree c a]

NodeWithCleanup (Maybe (String, Location)) c [Tree c a]

Leaf a

data Item a

Item is used to represent spec items internally. A spec item consists of:

a textual description of a desired behavior

an example for that behavior

additional meta information

Everything that is an instance of the Example type class can be used as an example, including QuickCheck properties, Hspec expectations and HUnit assertions.

Constructors

Item

Fields

itemRequirement :: String

Textual description of behavior

itemLocation :: Maybe Location

Source location of the spec item

itemIsParallelizable :: Maybe Bool

A flag that indicates whether it is safe to evaluate this spec item in parallel with other spec items

itemIsFocused :: Bool

A flag that indicates whether this spec item is focused

itemAnnotations :: Annotations

Arbitrary additional data that can be used by third-party extensions.

Since: 2.12.0

itemExample :: Params -> (ActionWith a -> IO ()) -> ProgressCallback -> IO Result

Example for behavior

specGroup :: HasCallStack => String -> [SpecTree a] -> SpecTree a

The specGroup function combines a list of specs into a larger spec.

specItem :: (HasCallStack, Example e) => String -> e -> SpecTree (Arg e)

The specItem function creates a spec item.

bimapTree :: (a -> b) -> (c -> d) -> Tree a c -> Tree b d

bimapForest :: (a -> b) -> (c -> d) -> [Tree a c] -> [Tree b d]

filterTree :: (a -> Bool) -> Tree c a -> Maybe (Tree c a)

filterForest :: (a -> Bool) -> [Tree c a] -> [Tree c a]

filterTreeWithLabels :: ([String] -> a -> Bool) -> Tree c a -> Maybe (Tree c a)

filterForestWithLabels :: ([String] -> a -> Bool) -> [Tree c a] -> [Tree c a]

pruneTree :: Tree c a -> Maybe (Tree c a)

pruneForest :: [Tree c a] -> [Tree c a]

location :: HasCallStack => Maybe Location

focusForest :: [SpecTree a] -> [SpecTree a]

Marks an entire spec forest as focused if nothing in it is already focused.

Re-exports

type HasCallStack = ?callStack :: CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.9.0.0

type Expectation = Assertion