Test.Hspec.Core.Extension.Config

Description

Warning: This API is experimental.

Types

data Config

Constructors

Config

Fields

configIgnoreConfigFile :: Bool

configDryRun :: Bool

configFocusedOnly :: Bool

configFailOnEmpty :: Bool

configFailOnFocused :: Bool

configFailOnPending :: Bool

configFailOnEmptyDescription :: Bool

configPrintSlowItems :: Maybe Int

configPrintCpuTime :: Bool

configFailFast :: Bool

configRandomize :: Bool

configSeed :: Maybe Integer

configQuickCheckSeed :: Maybe Integer

Deprecated: Use configSeed instead.

configFailureReport :: Maybe FilePath

configRerun :: Bool

configRerunAllOnSuccess :: Bool

configFilterPredicate :: Maybe (Path -> Bool)

configSkipPredicate :: Maybe (Path -> Bool)

configQuickCheckMaxSuccess :: Maybe Int

configQuickCheckMaxDiscardRatio :: Maybe Int

configQuickCheckMaxSize :: Maybe Int

configQuickCheckMaxShrinks :: Maybe Int

configSmallCheckDepth :: Maybe Int

configColorMode :: ColorMode

configUnicodeMode :: UnicodeMode

configDiff :: Bool

configDiffContext :: Maybe Int

configExternalDiff :: Maybe (Maybe Int -> String -> String -> IO ())

configPrettyPrint :: Bool

configPrettyPrintFunction :: Bool -> String -> String -> (String, String)

configFormatException :: SomeException -> String

Since: 2.11.5

configTimes :: Bool

configExpertMode :: Bool

Since: 2.11.2

configAvailableFormatters :: [(String, FormatConfig -> IO Format)]

Since: 2.9.0

configFormat :: Maybe (FormatConfig -> IO Format)

configFormatter :: Maybe Formatter

Deprecated: Use useFormatter instead.

configHtmlOutput :: Bool

configConcurrentJobs :: Maybe Int

configAnnotations :: Annotations

type Path = ([String], String)

A Path describes the location of a spec item within a spec tree.

It consists of a list of group descriptions and a requirement description.

Since: 2.0.0

data ColorMode

Constructors

ColorAuto

ColorNever

ColorAlways

data UnicodeMode

Constructors

UnicodeAuto

UnicodeNever

UnicodeAlways

Operations

setAnnotation :: Typeable value => value -> Config -> Config

getAnnotation :: Typeable value => Config -> Maybe value