Test.Hspec.Core.Formatters.V2

Description

This module contains formatters that can be used with hspecWith.

Formatters

silent :: Formatter

checks :: Formatter

specdoc :: Formatter

progress :: Formatter

failed_examples :: Formatter

Implementing a custom Formatter

A formatter is a set of actions. Each action is evaluated when a certain situation is encountered during a test run.

Actions live in the FormatM monad. It provides access to the runner state and primitives for appending to the generated report.

data Formatter

Constructors

Formatter

Fields

formatterStarted :: FormatM ()

formatterGroupStarted :: Path -> FormatM ()

formatterGroupDone :: Path -> FormatM ()

formatterProgress :: Path -> Progress -> FormatM ()

formatterItemStarted :: Path -> FormatM ()

formatterItemDone :: Path -> Item -> FormatM ()

formatterDone :: FormatM ()

data Item

Constructors

Item

Fields

itemLocation :: Maybe Location

itemDuration :: Seconds

itemInfo :: String

itemResult :: Result

data Result

Constructors

Success

Pending (Maybe Location) (Maybe String)

Failure (Maybe Location) FailureReason

data FailureReason

Constructors

NoReason

Reason String

ExpectedButGot (Maybe String) String String

Error (Maybe String) SomeException

data FormatM a

formatterToFormat :: Formatter -> FormatConfig -> IO Format

Accessing the runner state

getSuccessCount :: FormatM Int

Get the number of successful examples encountered so far.

getPendingCount :: FormatM Int

Get the number of pending examples encountered so far.

getFailCount :: FormatM Int

Get the number of failed examples encountered so far.

getTotalCount :: FormatM Int

Get the total number of examples encountered so far.

getExpectedTotalCount :: FormatM Int

Get the number of spec items that will have been encountered when this run completes (if it is not terminated early).

data FailureRecord

Constructors

FailureRecord

Fields

failureRecordLocation :: Maybe Location

failureRecordPath :: Path

failureRecordMessage :: FailureReason

getFailMessages :: FormatM [FailureRecord]

Get the list of accumulated failure messages.

usedSeed :: FormatM Integer

The random seed that is used for QuickCheck.

printTimes :: FormatM Bool

Return True if the user requested time reporting for individual spec items, False otherwise.

newtype Seconds

Constructors

Seconds Double

getCPUTime :: FormatM (Maybe Seconds)

Get the used CPU time since the test run has been started.

getRealTime :: FormatM Seconds

Get the passed real time since the test run has been started.

Appending to the generated report

write :: String -> FormatM ()

Append some output to the report.

writeLine :: String -> FormatM ()

The same as write, but adds a newline character.

writeTransient :: String -> FormatM ()

Dealing with colors

withInfoColor :: FormatM a -> FormatM a

Set output color to cyan, run given action, and finally restore the default color.

withSuccessColor :: FormatM a -> FormatM a

Set output color to green, run given action, and finally restore the default color.

withPendingColor :: FormatM a -> FormatM a

Set output color to yellow, run given action, and finally restore the default color.

withFailColor :: FormatM a -> FormatM a

Set output color to red, run given action, and finally restore the default color.

outputUnicode :: FormatM Bool

Return True if the user requested unicode output, False otherwise.

useDiff :: FormatM Bool

Return True if the user requested colorized diffs, False otherwise.

prettyPrint :: FormatM Bool

Return True if the user requested pretty diffs, False otherwise.

extraChunk :: String -> FormatM ()

Output given chunk in red.

missingChunk :: String -> FormatM ()

Output given chunk in green.

Helpers

formatLocation :: Location -> String

formatException :: SomeException -> String

The function formatException converts an exception to a string.

This is different from show. The type of the exception is included, e.g.:

>>> formatException (toException DivideByZero)
"ArithException (divide by zero)"

For IOExceptions the IOErrorType is included, as well.