Test.Hspec.Core.Hooks

Description

Documentation

before :: IO a -> SpecWith a -> Spec

Run a custom action before every spec item.

before_ :: IO () -> SpecWith a -> SpecWith a

Run a custom action before every spec item.

beforeWith :: (b -> IO a) -> SpecWith a -> SpecWith b

Run a custom action before every spec item.

beforeAll :: IO a -> SpecWith a -> Spec

Run a custom action before the first spec item.

beforeAll_ :: IO () -> SpecWith a -> SpecWith a

Run a custom action before the first spec item.

after :: ActionWith a -> SpecWith a -> SpecWith a

Run a custom action after every spec item.

after_ :: IO () -> SpecWith a -> SpecWith a

Run a custom action after every spec item.

afterAll :: ActionWith a -> SpecWith a -> SpecWith a

Run a custom action after the last spec item.

afterAll_ :: IO () -> SpecWith a -> SpecWith a

Run a custom action after the last spec item.

around :: (ActionWith a -> IO ()) -> SpecWith a -> Spec

Run a custom action before and/or after every spec item.

around_ :: (IO () -> IO ()) -> SpecWith a -> SpecWith a

Run a custom action before and/or after every spec item.

aroundWith :: (ActionWith a -> ActionWith b) -> SpecWith a -> SpecWith b

Run a custom action before and/or after every spec item.